+++ components/app/crates/tt-rs-app/src/app/callbacks/mod.rs
@@ -7,7 +7,7 @@ use std::cell::RefCell;
 use std::rc::Rc;
 use tt_rs_core::WidgetId;
 use tt_rs_drag::{CopySourceClickEvent, DragEndEvent, DragStartEvent, DropEvent, Position};
-use tt_rs_ui::UserLevel;
+use tt_rs_ui::{SaveFormData, UserLevel};
 use yew::prelude::*;
 
 use crate::state::AppState;
@@ -22,12 +22,21 @@ pub struct Callbacks {
     pub on_copy_source_click: Callback<CopySourceClickEvent>,
     pub on_move: Callback<(WidgetId, Position)>,
     pub on_drop: Callback<DropEvent>,
+    // Workspace callbacks
+    pub on_workspace_open: Callback<()>,
+    pub on_workspace_close: Callback<()>,
+    pub on_workspace_save: Callback<SaveFormData>,
+    pub on_workspace_load: Callback<String>,
+    pub on_workspace_delete: Callback<String>,
+    pub on_workspace_export: Callback<String>,
+    pub on_workspace_import: Callback<web_sys::File>,
 }
 
 pub fn create_callbacks(
     state: UseStateHandle<AppState>,
     help_open: UseStateHandle<bool>,
     user_level: UseStateHandle<UserLevel>,
+    workspace_open: UseStateHandle<bool>,
     dragged_box_id: Rc<RefCell<Option<WidgetId>>>,
     pending_new_box: Rc<RefCell<Option<usize>>>,
 ) -> Callbacks {
@@ -50,5 +59,34 @@ pub fn create_callbacks(
         on_copy_source_click: widget_handlers::create_copy_source(state.clone()),
         on_move: widget_handlers::create_move(state.clone()),
         on_drop: widget_handlers::create_widget_drop(state),
+        // Workspace callbacks - stubs for now, will be implemented with storage
+        on_workspace_open: {
+            let w = workspace_open.clone();
+            Callback::from(move |_| w.set(true))
+        },
+        on_workspace_close: {
+            let w = workspace_open;
+            Callback::from(move |_| w.set(false))
+        },
+        on_workspace_save: Callback::from(|data: SaveFormData| {
+            log::info!("Save workspace: {} - {}", data.name, data.description);
+            // TODO: Implement actual save with storage backend
+        }),
+        on_workspace_load: Callback::from(|id: String| {
+            log::info!("Load workspace: {}", id);
+            // TODO: Implement actual load with storage backend
+        }),
+        on_workspace_delete: Callback::from(|id: String| {
+            log::info!("Delete workspace: {}", id);
+            // TODO: Implement actual delete with storage backend
+        }),
+        on_workspace_export: Callback::from(|id: String| {
+            log::info!("Export workspace: {}", id);
+            // TODO: Implement actual export to file
+        }),
+        on_workspace_import: Callback::from(|_file: web_sys::File| {
+            log::info!("Import workspace from file");
+            // TODO: Implement actual import from file
+        }),
     }
 }